1) The deepest architectural change is a strong move towards tiling    everything.  GraphicsNodeRable now returns a GraphicsNodeRed that    is a tiled representation of the entire GraphicsNode.  This is then    rendered on demand.  For a variety of reasons this can be slower    than the pre-existing code in some cases, and _much_ faster in    others (panning around documents that use filters for example).
   This still needs more work, but what I'm checking in is currently
   fairly comparable to the pre-existing code performance wise and
   still offers significant low-hanging fruit for further improvement.

   One thing that I plan on doing very soon is modifying the renderer
   To use the caching code, this will make panning in large documents
   fairly snappy (we might even be able to pull off interactive redraw
   in most cases, based on a few tests).

2) Pattern/feTile this code has been completely reworked.  The new
   code appears to be faster in a number of cases, and is comparable
   in the rest.

3) feDisplacementMap this has also been heavily reworked.  The default
   implementation now uses Bi-Linear interpolation.
   samples/tests/feDisplacementMap.svg looks a little better, but the
   BE test looks sweet (no more jaggies)!

   The Anti-aliased code is a bit slower than the original code, (it
   still has room to improve with the techniques I used for the
   Nearest Neighbor code) I have left in a Nearest Neighbor version
   that is at least 2x as fast as the original code.  I also reworked
   to code so it only generates the region of output needed (this is
   very close to a tiled implementation and should carry over nicely).

   Eventually which method is used should be controlled by a hint.

4) 'createGraphics'.  I have added a createGraphics method to
   GraphicsUtils, along with a number of 'getDevice*' calls.  These
   methods work together to ensure that information about the
   destination is available when using a Graphics2D.  I have replaced
   most usages of BufferedImage.createGraphics with the equivalent
   call to GraphicsUtils.createGraphics. In some instances where the
   graphics was only used for very short lived work (g.dispose was
   called shortly afterwords) I didn't bother to introduce the
   dependency on GraphicsUtils.

   I think this will help to squash the constantly re-occuring bugs
   where things are not properly converted to/from linear sRGB.


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@199434 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/apps/regard/Main.java
@@ -134,21 +134,27 @@ public class Main {
     public static void reset() {
         File dir = new File(getRefDirectory());
         File [] images = dir.listFiles();
-        for (int i=0; i < images.length; ++i) {
-            display("Deleting reference image "+images[i].getName());
-            images[i].delete();
+        if (images != null) {
+            for (int i=0; i < images.length; ++i) {
+                display("Deleting reference image "+images[i].getName());
+                images[i].delete();
+            }
         }
         dir = new File(getNewDirectory());
         images = dir.listFiles();
-        for (int i=0; i < images.length; ++i) {
-            display("Deleting new image "+images[i].getName());
-            images[i].delete();
+        if (images != null) {
+            for (int i=0; i < images.length; ++i) {
+                display("Deleting new image "+images[i].getName());
+                images[i].delete();
+            }
         }
         dir = new File(getDiffDirectory());
         images = dir.listFiles();
-        for (int i=0; i < images.length; ++i) {
-            display("Deleting diff image "+images[i].getName());
-            images[i].delete();
+        if (images != null) {
+            for (int i=0; i < images.length; ++i) {
+                display("Deleting diff image "+images[i].getName());
+                images[i].delete();
+            }
         }
         exit(0);
     }
@@ -165,12 +171,14 @@ public class Main {
             error("regard is not initialized. use -init first");
         }
         File ff = new File(f, REGARD_REF_DIRECTORY_NAME);
-        if (ff.list().length != 0) {
+        String [] files = ff.list();
+        if ((files!=null) && (files.length != 0)) {
             error("The reference directory "+ff.getAbsolutePath()+
                   " is not empty. Use -reset first.");
             exit(3);
         }
         generateImages(getRefDirectory(), "reference image for");
+        exit(0);
     }
 
     /**
@@ -185,6 +193,7 @@ public class Main {
             error("regard is not initialized. use -init first");
         }
         generateImages(getNewDirectory(), "new image for");
+        exit(0);
     }
 
     static void generateImages(String outputDirectory, String desc) {
@@ -333,10 +342,6 @@ public class Main {
             BufferedImage bfDiff = transcoder.createImage(2*bfRef.getWidth(),
                                                           2*bfRef.getHeight());
 
-            Graphics2D g = bfDiff.createGraphics();
-            // g.setPaint(transcoder.getBackgroundPaint());
-            // g.fillRect(0, 0, bfDiff.getWidth(), bfDiff.getHeight());
-            g.dispose();
             display(String.valueOf(i+1) + ". " + "Creating the difference image file of " + refImg.getName());
 
             boolean difference =
